# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pharmacy POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medicines:
    get:
      operationId: listMedicines
      tags: [medicines]
      summary: List medicines
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search by name, generic name, or brand
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: manufacturer
          in: query
          schema:
            type: string
        - name: requires_prescription
          in: query
          schema:
            type: boolean
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: sort_by
          in: query
          schema:
            type: string
        - name: sort_order
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated medicines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicineListResponse"
    post:
      operationId: createMedicine
      tags: [medicines]
      summary: Create medicine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
  /medicines/search:
    get:
      operationId: searchMedicines
      tags: [medicines]
      summary: Search medicines
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: barcode
          in: query
          schema:
            type: string
        - name: branch_id
          in: query
          schema:
            type: string
        - name: in_stock_only
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicine"
  /medicines/low-stock:
    get:
      operationId: getLowStockMedicines
      tags: [medicines]
      summary: Get low-stock medicines
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
        - name: threshold
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Low stock medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LowStockMedicine"
  /medicines/expiring-soon:
    get:
      operationId: getExpiringMedicines
      tags: [medicines]
      summary: Get expiring-soon medicines
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 90
        - name: branch_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expiring medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpiringMedicine"
...
  /medicines/{medicine_id}:
    get:
      operationId: getMedicine
      tags: [medicines]
      summary: Get medicine detail
      parameters:
        - name: medicine_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Medicine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    put:
      operationId: updateMedicine
      tags: [medicines]
      summary: Update medicine
      parameters:
        - name: medicine_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    delete:
      operationId: deactivateMedicine
      tags: [medicines]
      summary: Deactivate medicine
      parameters:
        - name: medicine_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── INVENTORY ──────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
      required:
        - page
        - per_page
        - total
        - pages
        - has_next
        - has_prev

    # ── AUTH ──────────────────────────────────────────────────
...
    Medicine:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        generic_name:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        manufacturer:
          type: ["string", "null"]
        strength:
          type: ["string", "null"]
        form:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        pack_size:
          type: ["integer", "null"]
        barcode:
          type: ["string", "null"]
        mrp:
          type: number
        requires_prescription:
          type: boolean
        controlled_substance_schedule:
          type: ["string", "null"]
        is_essential_medicine:
          type: boolean
        is_active:
          type: boolean
        min_stock_level:
          type: ["integer", "null"]
        reorder_point:
          type: ["integer", "null"]
        storage_conditions:
          type: ["string", "null"]
        current_stock:
          type: ["integer", "null"]
        created_at:
          type: string
      required:
        - id
        - name
        - mrp
        - requires_prescription
        - is_essential_medicine
        - is_active
        - created_at
    MedicineInput:
      type: object
      properties:
        name:
          type: string
        generic_name:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        manufacturer:
          type: ["string", "null"]
        strength:
          type: ["string", "null"]
        form:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        pack_size:
          type: ["integer", "null"]
        barcode:
          type: ["string", "null"]
        mrp:
          type: number
        requires_prescription:
          type: boolean
        is_essential_medicine:
          type: boolean
        min_stock_level:
          type: ["integer", "null"]
        reorder_point:
          type: ["integer", "null"]
        storage_conditions:
          type: ["string", "null"]
        composition:
          type: ["string", "null"]
        drap_registration_no:
          type: ["string", "null"]
      required:
        - name
        - mrp
        - requires_prescription
    MedicineUpdate:
      type: object
      properties:
        name:
          type: string
        generic_name:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        mrp:
          type: number
        requires_prescription:
          type: boolean
        is_active:
          type: boolean
        min_stock_level:
          type: ["integer", "null"]
        reorder_point:
          type: ["integer", "null"]
    MedicineListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Medicine"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    LowStockMedicine:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: ["string", "null"]
        current_stock:
          type: integer
        min_stock_level:
          type: integer
        reorder_point:
          type: integer
        branch_id:
          type: string
        branch_name:
          type: string
      required:
        - id
        - name
        - current_stock
        - min_stock_level
        - reorder_point
    ExpiringMedicine:
      type: object
      properties:
        batch_id:
          type: string
        batch_number:
          type: string
        medicine_id:
          type: string
        medicine_name:
          type: string
        expiry_date:
          type: string
        days_until_expiry:
          type: integer
        quantity:
          type: integer
        branch_name:
          type: string
        urgency:
          type: string
      required:
        - batch_id
        - batch_number
        - medicine_id
        - medicine_name
        - expiry_date
        - days_until_expiry
        - quantity
...